Update README.md
This project is a front-end clone of the Amazon website, created using HTML and CSS. The site mimics the layout and design of Amazon's platform, focusing on desktop screens. Currently, the project is not responsive and is best viewed on desktop browsers. Future updates will include responsiveness and added interactivity.